Crate pandemonium

FreebSD system libraries binding

Technically should seprate create per lib, but right now it's easy that way. Also, it does pretty much nothing and mostly used by me to learn rust.:w

Modules

errors

jail
libjail binding for Rust.

sysinfo
FreeBSD's libc binding for rust.