Module pandemonium::jail::ffi

Reexports

pub use self::operation::*;

Modules

operation

Structs

NameIter
Low leven iterator for jails. The idea is to use this iterator for every other iterator since the only thing it returns is jail name.

Functions

jid_iter
Nothing more than a shortcut to get iterator for jails.

name_for_jid