Struct pandemonium::jail::ffi::operation::JailOperation

pub struct JailOperation {
    // some fields omitted
}

Safe interface for operation with jails

Methods

impl JailOperation

fn new(name: String) -> Result<JailOperation, Error>

Proper way of setting up Operation

fn get(&mut self) -> Result<Vec<Option<String>>, Error>

executes read operation on jail and returns Vector of values for requested keys. Values returned in the same order they were requested

fn add_jailparams(&mut self, params: Vec<ParamKey>) -> Result<(), Error>

fn add_jaiparams_with_values(&mut self, params: Vec<(ParamKey, String)>) -> Result<(), Error>

fn add_jailparam_with_value(&mut self, key: &ParamKey, value: &String) -> Result<(), Error>

Use this if you want set value to jail

fn add_jailparam(&mut self, key: &ParamKey) -> Result<(), Error>

Use this if you want get value from jail