Struct pandemonium::zfs::types::Dataset

pub struct Dataset {
    pub name: String,
    pub mounted: bool,
    pub available: String,
    pub used: String,
    pub compression: String,
    pub mount_point: String,
}

Fields

name

mounted

available

used

compression

mount_point

Methods

impl Dataset

fn from_line(line: &str) -> Dataset

Contructor to make Dataset object from output of zfs list command

fn from_stdout(stdout: String) -> Vec<Dataset>

Contructor to make vector of Datasets from zfs list command